+++ CHANGELOG.md
@@ -15,6 +15,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 - Support the `disabled` prop on the `CentsField` component.
 - Support the `disabled` prop on the `DateTimeField` component.
 - Support the `disabled` prop on the `FileField` component.
+- Support the `disabled` prop on the `RadioField` component.
+- Support the `disabled` prop on the `TextField` component.
 
 ## [16.3.0] - 2020-01-07